orders/src/events/listener/expiration-complete-listener.ts
import { Listener, ExpirationCompleteEvent, Subjects, OrderStatus } from '@mh132001tickets/common'
import { Message } from 'node-nats-streaming';
import { queueGroupName } from './queue-group-name';
import { Order } from '../../models/order';
import { OrderCancelledPublisher } from '../publisher/order-cancelled-publisher';

export class ExpirationCompleteListener extends Listener<ExpirationCompleteEvent>{
    subject: Subjects.ExpirationComplete = Subjects.ExpirationComplete;
    queueGroupName = queueGroupName;

    async onMessage(data: ExpirationCompleteEvent['data'], msg: Message) {
        const order = await Order.findOne({id: data.orderId})

        if (!order){
            throw new Error('Order not found!');
        }

        order.status = OrderStatus.Cancelled;
        await order.save();
        await new OrderCancelledPublisher(this.client).publish({
            id: order.id,
            version: order.version,
            ticket: {
                id: order.ticket.id,
            }
        })
        
        msg.ack();
    }
}